+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+* Add support for `user_id` query string parameter when loading teams
 
 ## [v3.6.0](https://github.com/buildkite/go-buildkite/compare/v3.5.0...v3.6.0) (2023-09-18)
 * SUP-1353: Clusters integration [#154](https://github.com/buildkite/go-buildkite/pull/154) ([james2791](https://github.com/james2791))